This branch hasn't been updated with changes from HEAD for a looooong time... Here goes.
This also include a currently half-done attempt to convert to using size_t
appropriately.

+++ crypto/hmac/hmac.c
@@ -80,7 +80,7 @@ void HMAC_Init_ex(HMAC_CTX *ctx, const void *key, size_t len,
 		{
 		reset=1;
 		j=EVP_MD_block_size(md);
-		OPENSSL_assert(j <= sizeof ctx->key);
+		OPENSSL_assert(j <= (int)sizeof(ctx->key));
 		if (j < len)
 			{
 			EVP_DigestInit_ex(&ctx->md_ctx,md, impl);
@@ -90,7 +90,7 @@ void HMAC_Init_ex(HMAC_CTX *ctx, const void *key, size_t len,
 			}
 		else
 			{
-			OPENSSL_assert(len <= sizeof ctx->key);
+			OPENSSL_assert(len <= (int)sizeof(ctx->key));
 			memcpy(ctx->key,key,len);
 			ctx->key_length=len;
 			}
